fix: keep pending local echo helper css-free
